+++ R/RcppExports.R
@@ -1659,6 +1659,91 @@ vsi_rename <- function(oldpath, newpath) {
     invisible(.Call(`_gdalraster_vsi_rename`, oldpath, newpath))
 }
 
+#' Return the list of virtual file system handlers currently registered
+#'
+#' `vsi_get_fs_prefixes()` returns the list of prefixes for virtual file
+#' system handlers currently registered (e.g., `"/vsimem/"`, `"/vsicurl/"`,
+#' etc). Wrapper for `VSIGetFileSystemsPrefixes()` in the GDAL API.
+#'
+#' @returns Character vector containing prefixes of the virtual file system
+#' handlers.
+#'
+#' @seealso
+#' [vsi_get_fs_options()]
+#'
+#' \url{https://gdal.org/user/virtual_file_systems.html}
+#'
+#' @examples
+#' vsi_get_fs_prefixes()
+vsi_get_fs_prefixes <- function() {
+    .Call(`_gdalraster_vsi_get_fs_prefixes`)
+}
+
+#' Return the list of options associated with a virtual file system handler
+#' as a serialized XML string.
+#'
+#' Called from and documented in R/gdal_helpers.R
+#' @noRd
+.vsi_get_fs_options <- function(filename) {
+    .Call(`_gdalraster__vsi_get_fs_options`, filename)
+}
+
+#' Return whether the filesystem supports sequential write
+#'
+#' `vsi_supports_seq_write()` returns whether the filesystem supports
+#' sequential write.
+#' Wrapper for `VSISupportsSequentialWrite()` in the GDAL API.
+#'
+#' @param filename Character string. The path of the filesystem object to be
+#' tested.
+#' @param allow_local_tmpfile Logical scalar. `TRUE` if the filesystem is
+#' allowed to use a local temporary file before uploading to the target
+#' location.
+#' @returns Logical scalar. `TRUE` if sequential write is supported.
+#'
+#' @note
+#' The location GDAL uses for temporary files can be forced via the
+#' `CPL_TMPDIR` configuration option.
+#'
+#' @seealso
+#' [vsi_supports_rnd_write()]
+#'
+#' @examples
+#' # Requires GDAL >= 3.6
+#' if (as.integer(gdal_version()[2]) >= 3060000)
+#'   vsi_supports_seq_write("/vsimem/test-mem-file.gpkg", TRUE)
+vsi_supports_seq_write <- function(filename, allow_local_tmpfile) {
+    .Call(`_gdalraster_vsi_supports_seq_write`, filename, allow_local_tmpfile)
+}
+
+#' Return whether the filesystem supports random write
+#'
+#' `vsi_supports_rnd_write()` returns whether the filesystem supports
+#' random write.
+#' Wrapper for `VSISupportsRandomWrite()` in the GDAL API.
+#'
+#' @param filename Character string. The path of the filesystem object to be
+#' tested.
+#' @param allow_local_tmpfile Logical scalar. `TRUE` if the filesystem is
+#' allowed to use a local temporary file before uploading to the target
+#' location.
+#' @returns Logical scalar. `TRUE` if random write is supported.
+#'
+#' @note
+#' The location GDAL uses for temporary files can be forced via the
+#' `CPL_TMPDIR` configuration option.
+#'
+#' @seealso
+#' [vsi_supports_seq_write()]
+#'
+#' @examples
+#' # Requires GDAL >= 3.6
+#' if (as.integer(gdal_version()[2]) >= 3060000)
+#'   vsi_supports_rnd_write("/vsimem/test-mem-file.gpkg", TRUE)
+vsi_supports_rnd_write <- function(filename, allow_local_tmpfile) {
+    .Call(`_gdalraster_vsi_supports_rnd_write`, filename, allow_local_tmpfile)
+}
+
 #' @noRd
 NULL