+++ src/gdal_vsi.cpp
@@ -630,3 +630,143 @@ int vsi_rename(Rcpp::CharacterVector oldpath, Rcpp::CharacterVector newpath) {
 	return VSIRename(oldpath_in.c_str(), newpath_in.c_str());
 }
 
+
+//' Return the list of virtual file system handlers currently registered
+//'
+//' `vsi_get_fs_prefixes()` returns the list of prefixes for virtual file
+//' system handlers currently registered (e.g., `"/vsimem/"`, `"/vsicurl/"`,
+//' etc). Wrapper for `VSIGetFileSystemsPrefixes()` in the GDAL API.
+//'
+//' @returns Character vector containing prefixes of the virtual file system
+//' handlers.
+//'
+//' @seealso
+//' [vsi_get_fs_options()]
+//'
+//' \url{https://gdal.org/user/virtual_file_systems.html}
+//'
+//' @examples
+//' vsi_get_fs_prefixes()
+// [[Rcpp::export()]]
+Rcpp::CharacterVector vsi_get_fs_prefixes() {
+
+	char **papszPrefixes;
+	papszPrefixes = VSIGetFileSystemsPrefixes();
+	
+	int items = CSLCount(papszPrefixes);
+	if (items > 0) {
+		Rcpp::CharacterVector prefixes(items);
+		for (int i=0; i < items; ++i) {
+			prefixes(i) = papszPrefixes[i];
+		}
+		CSLDestroy(papszPrefixes);
+		return prefixes;
+	}
+	else {
+		CSLDestroy(papszPrefixes);
+		return "";	
+	}
+}
+
+
+//' Return the list of options associated with a virtual file system handler
+//' as a serialized XML string.
+//'
+//' Called from and documented in R/gdal_helpers.R
+//' @noRd
+// [[Rcpp::export(name = ".vsi_get_fs_options")]]
+std::string _vsi_get_fs_options(Rcpp::CharacterVector filename) {
+
+	std::string filename_in;
+	filename_in = Rcpp::as<std::string>(_check_gdal_filename(filename));
+	if (VSIGetFileSystemOptions(filename_in.c_str()) != NULL)
+		return VSIGetFileSystemOptions(filename_in.c_str());
+	else
+		return "";
+}
+
+
+//' Return whether the filesystem supports sequential write
+//'
+//' `vsi_supports_seq_write()` returns whether the filesystem supports
+//' sequential write.
+//' Wrapper for `VSISupportsSequentialWrite()` in the GDAL API.
+//'
+//' @param filename Character string. The path of the filesystem object to be
+//' tested.
+//' @param allow_local_tmpfile Logical scalar. `TRUE` if the filesystem is
+//' allowed to use a local temporary file before uploading to the target
+//' location.
+//' @returns Logical scalar. `TRUE` if sequential write is supported.
+//'
+//' @note
+//' The location GDAL uses for temporary files can be forced via the
+//' `CPL_TMPDIR` configuration option.
+//'
+//' @seealso
+//' [vsi_supports_rnd_write()]
+//'
+//' @examples
+//' # Requires GDAL >= 3.6
+//' if (as.integer(gdal_version()[2]) >= 3060000)
+//'   vsi_supports_seq_write("/vsimem/test-mem-file.gpkg", TRUE)
+// [[Rcpp::export()]]
+bool vsi_supports_seq_write(Rcpp::CharacterVector filename,
+		bool allow_local_tmpfile) {
+
+#if GDAL_VERSION_NUM < 3060000
+	Rcpp::stop("vsi_supports_seq_write() requires GDAL >= 3.6.");
+
+#else
+	std::string filename_in;
+	filename_in = Rcpp::as<std::string>(_check_gdal_filename(filename));
+	if (VSISupportsSequentialWrite(filename_in.c_str(), allow_local_tmpfile))
+		return true;
+	else
+		return false;
+		
+#endif
+}
+
+
+//' Return whether the filesystem supports random write
+//'
+//' `vsi_supports_rnd_write()` returns whether the filesystem supports
+//' random write.
+//' Wrapper for `VSISupportsRandomWrite()` in the GDAL API.
+//'
+//' @param filename Character string. The path of the filesystem object to be
+//' tested.
+//' @param allow_local_tmpfile Logical scalar. `TRUE` if the filesystem is
+//' allowed to use a local temporary file before uploading to the target
+//' location.
+//' @returns Logical scalar. `TRUE` if random write is supported.
+//'
+//' @note
+//' The location GDAL uses for temporary files can be forced via the
+//' `CPL_TMPDIR` configuration option.
+//'
+//' @seealso
+//' [vsi_supports_seq_write()]
+//'
+//' @examples
+//' # Requires GDAL >= 3.6
+//' if (as.integer(gdal_version()[2]) >= 3060000)
+//'   vsi_supports_rnd_write("/vsimem/test-mem-file.gpkg", TRUE)
+// [[Rcpp::export()]]
+bool vsi_supports_rnd_write(Rcpp::CharacterVector filename,
+		bool allow_local_tmpfile) {
+
+#if GDAL_VERSION_NUM < 3060000
+	Rcpp::stop("vsi_supports_rnd_write() requires GDAL >= 3.6.");
+
+#else
+	std::string filename_in;
+	filename_in = Rcpp::as<std::string>(_check_gdal_filename(filename));
+	if (VSISupportsRandomWrite(filename_in.c_str(), allow_local_tmpfile))
+		return true;
+	else
+		return false;
+		
+#endif
+}